+++ fiddles/angular2-cli/fiddle-0023-EntAgGridRowSelection/src/app/ag-tree-grid/enums/ag-tree-grid-cmds.ts
@@ -1,0 +1,13 @@
+export enum AgTreeGridCmds {
+  ShowAll = 1,
+  ShowSelected = 2,
+  ClearSelected = 3,
+  Reload = 4,
+  ExpandAll = 5,
+  CollapseAll = 6,
+  SelectAll = 7,
+  Export = 8,
+  showLoader = 9,
+  hideLoader = 10,
+  invalidate =11
+}