+++ fiddles/angular2-cli/fiddle-0023-EntAgGridRowSelection/src/app/ag-tree-grid/interfaces/ag-tree-grid-options.ts
@@ -1,0 +1,11 @@
+
+export interface AgTreeGridOptions {
+
+  autoGroupColumnDef?: any;
+
+  columnDefs?: any[];
+
+  rowData?: any[];
+
+
+}